chore: Update npm dependencies and fix prettier configuration
This commit updates the npm dependencies to their latest versions and fixes the prettier configuration in the `prettier.config.mjs` file. The `semi` option is changed from `true` to `false`, ensuring that semicolons are not automatically inserted in the code. Additionally, the `@typescript-eslint/strict` plugin is commented out in the `.eslintrc.json` file.

Recent user commits:
- fix: added security by changing how db actions get executed
- chore: Update npm dependencies and add Next UI components
- Merge pull request #55 from wizelineacademy/H5-IntegrationP1P2
- Merge branch 'H010-UserHistory' into H5-IntegrationP1P2
- feat: Update navigation to history page and redirect to phase 3 dashboard
- Merge pull request #54 from wizelineacademy/H008-Front
- Merge branch 'H5-IntegrationP1P2' into H008-Front
- feat: Update navigation to history page and redirect to phase 3 dashboard
- Linting Fixes